services/users-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pymongo==4.6.0
kafka-python==2.0.2
aiokafka==0.10.0
python-dotenv==1.0.0
motor==3.3.2
email-validator==2.1.0
